Grand VFS restructure
Now VFS layer is responsible for vnodes, and ZFS can only call
vnode_recycle to terminate a vnode early (see zfs_remove). Normal
situations, vnodes are left as iocount,usecount = 0,0. Then vnode and
znode is ready to be used immediately in IRP_MJ_CREATE.

VFS layer keeps a count of number of vnodes, and once over
vnode_max, it will start reclaiming idle nodes. It calls
zfs_vnop_reclaim to let ZFS know to flush and close the file, before
the vnode is freed.

If case-insensitive pools are used, projects can be loaded into
Devstudio, but compiling fails.

+++ ZFSin/zfs/include/sys/zfs_znode.h
@@ -228,7 +228,7 @@ typedef struct znode {
     krwlock_t   z_map_lock;     /* page map lock */
 
 #ifdef _WIN32
-    uint32_t    z_vid;  /* OSX vnode_vid */
+    uint64_t    z_vid;  /* OSX vnode_vid */
 	boolean_t   z_fastpath;
 
 	uint32_t    z_name_len;    // Allocated size of z_name_cache
